refactor: FilterableRestaurantTable component 추가
